Moves sourcemap tests to index.test.js
